Move store info to app directory
